Split macro to files to improve order of code
